cleanup GgmlLoader and move primitives to MathPrimitives
